clean up struct array module
