+++ vortex-array/src/arrays/struct_/vtable/array.rs
@@ -1,0 +1,22 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use vortex_dtype::DType;
+
+use crate::arrays::struct_::{StructArray, StructVTable};
+use crate::stats::StatsSetRef;
+use crate::vtable::ArrayVTable;
+
+impl ArrayVTable<StructVTable> for StructVTable {
+    fn len(array: &StructArray) -> usize {
+        array.len
+    }
+
+    fn dtype(array: &StructArray) -> &DType {
+        &array.dtype
+    }
+
+    fn stats(array: &StructArray) -> StatsSetRef<'_> {
+        array.stats_set.to_ref(array.as_ref())
+    }
+}
