+++ vortex-array/src/arrays/struct_/vtable/pipeline.rs
